% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDamage.R
\docType{data}
\name{DataDamage}
\alias{DataDamage}
\title{Data damage}
\format{
A data frame with sources of solution, one in each column.
}
\usage{
data(DataDamage)
}
\description{
Example with data from data damage .
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
