% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLossSource.R
\docType{data}
\name{DataLossSource}
\alias{DataLossSource}
\title{Loss sources data}
\format{
A data frame with sources of loss, one in each column.
}
\usage{
data(DataLossSource)
}
\description{
Example with data from loss sources .
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
