% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSolutionSource.R
\docType{data}
\name{DataSolutionSource}
\alias{DataSolutionSource}
\title{Solution sources data}
\format{
A data frame with sources of solution, one in each column.
}
\usage{
data(DataSolutionSource)
}
\description{
Example with data from solution sources .
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
