% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDefoliation.R
\docType{data}
\name{DataDefoliation}
\alias{DataDefoliation}
\title{Data defoliation}
\format{
A data frame with data defoliation.
}
\usage{
data(DataDefoliation)
}
\description{
Example with data from data defoliation .
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
