% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLossSource.R
\docType{data}
\name{DataLossSource}
\alias{DataLossSource}
\title{Loss sources data}
\format{
A data frame with sources of loss, one in each column.
}
\usage{
data(DataLossSource)
}
\description{
An example with data from loss sources.
}
\references{
Demolin-Leite, G. L. (2024), Percentage of importance indice-production unknown:
loss and solution sources identification on system. Brazilian Journal of Biology 84, e253218.
 <https://doi.org/10.1590/1519-6984.253218>.
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
