% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSolutionSource.R
\docType{data}
\name{DataSolutionSource}
\alias{DataSolutionSource}
\title{Solution sources data}
\format{
A data frame with sources of solution, one in each column.
}
\usage{
data(DataSolutionSource)
}
\description{
An example with data from solution sources .
}
\references{
Demolin-Leite, G. L. (2024), Percentage of importance indice-production unknown:
loss and solution sources identification on system. Brazilian Journal of Biology 84, e253218.
 <https://doi.org/10.1590/1519-6984.253218>.
}
\author{
Germano Leao Demolin Leite : \email{germano.demolin@gmail.com}

Alcinei Mistico Azevedo : \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
